fix for bar width/position calculation
git-svn-id: https://efmsvn2.nrel.gov/wex/svn/trunk@474 e21c72fa-1056-4880-a922-3284d9a13290

+++ include/wex/plot/plbarplot.h
@@ -52,7 +52,7 @@ public:
 protected:
 
 	double CalcYPos(double x) const;
-	double CalcXPos(double x, const wxPLDeviceMapping &map, int dispwidth);
+	double CalcXPos(double x, const wxPLDeviceMapping &map, double dispwidth);
 	double CalcDispBarWidth( const wxPLDeviceMapping &map );
 	
 	wxPLBarPlot *m_stackedOn;
